Modularizes js build utils. Adds a test.
