+++ webpack.config.js
@@ -1,9 +1,8 @@
 var path = require("path");
 var webpack = require('webpack');
-var BundleTracker = require('webpack-bundle-tracker');
-var iniParser = require('ini-parser');
-var fs = require('fs');
+
 var logging = require('./assets/js/logging');
+var iniFiles = require('./assets/js/ini_files');
 
 var bundles = [];
 
@@ -12,48 +11,8 @@ var js_source_dirs = [
     "kolibri/plugins"
 ];
 
-var parseBundleIni = function(iniFile) {
-    var data = iniParser.parse(fs.readFileSync(iniFile, 'utf-8'));
-    var bundle_data = {};
-    for (var key in data) {
-        if (typeof data[key].entry_file !== "undefined") {
-            bundle_data[key] = path.join(path.dirname(iniFile), data[key].entry_file);
-        } else {
-            logging.error(iniFile + " file is misconfigured, missing 'entry_file' parameter for bundle " + key);
-        }
-    }
-
-    var module_name = path.relative(__dirname, path.dirname(iniFile)).replace(/\//g, ".");
-
-    if (Object.keys(bundle_data).length > 0) {
-        bundles.push({
-            context: __dirname,
-            entry: bundle_data,
-            output: {
-                path: path.join(path.dirname(iniFile), "static", module_name),
-                filename: "[name]-[hash].js"
-            },
-            plugins: [
-                new BundleTracker({path: path.dirname(iniFile), filename: "bundle_catalog.json"})
-            ]
-        });
-    }
-};
-
-var recurseBundleIni = function(directory) {
-    var files = fs.readdirSync(directory);
-    files.forEach(function(file){
-        var stats = fs.statSync(path.join(directory, file));
-        if (stats.isDirectory()) {
-            recurseBundleIni(path.join(directory, file));
-        } else if (file.indexOf("bundles.ini") > -1) {
-            parseBundleIni(path.join(directory, file));
-        }
-    });
-};
-
 for (var i = 0; i < js_source_dirs.length; i++) {
-    recurseBundleIni(path.join(__dirname, js_source_dirs[i]));
+    iniFiles.recurseBundleIni(path.join(__dirname, js_source_dirs[i]), bundles);
 }
 
 module.exports = bundles;
